Update docs and rename files
