+++ packages/integration-tests/Executions.js
@@ -1,5 +1,12 @@
 'use strict';
 
+/**
+ * @module Executions
+ *
+ * @example
+ * const Executions = require('@cumulus/integration-test/Executions');
+ */
+
 const executionsApi = require('@cumulus/api-client/executions');
 const get = require('lodash/get');
 const isNil = require('lodash/isNil');
@@ -9,13 +16,15 @@ const pRetry = require('p-retry');
 /**
  * Find the execution ARN matching the `matcher` function
  *
- * @param {string} prefix - the prefix configured for the stack
- * @param {Function} matcher - a predicate function that takes an execution and
- *   determines if this is the execution that is being searched for
+ * @param {string} prefix - the name of the Cumulus stack
+ * @param {Function} matcher - a predicate function that takes an execution and determines if this
+ * is the execution that is being searched for
  * @param {Object} [options]
- * @param {integer} [options.timeout] - the number of seconds to wait for a
- *   matching execution to be found
- * @returns {Promise<Object>} the matching execution
+ * @param {integer} [options.timeout=0] - the number of seconds to wait for a matching execution
+ * to be found
+ * @returns {Promise<string>} the ARN of the matching execution
+ *
+ * @alias module:Executions
  */
 const findExecutionArn = async (prefix, matcher, options = { timeout: 0 }) =>
   pRetry(
@@ -41,19 +50,21 @@ const findExecutionArn = async (prefix, matcher, options = { timeout: 0 }) =>
   );
 
 /**
- * Wait for an execution to be completed and return it
+ * Wait for an execution status to be `completed` and return the execution
  *
  * @param {Object} params
- * @param {string} params.prefix - the prefix configured for the stack
- * @param {string} params.arn - an execution ARN
- * @param {Function} params.callback - an async function to invoke the api
- *   lambda that takes a prefix / user payload. Defaults to
- *   cumulusApiClient.invokeApifunction to invoke the api lambda
- * @param {integer} params.timeout - the number of seconds to wait for the
- *   execution to reach a terminal state. Defaults to 30
- * @returns {Promise<Object>} the completed execution
+ * @param {string} params.prefix - the name of the Cumulus stack
+ * @param {string} params.arn - the execution ARN to fetch
+ * @param {Function} [params.callback=cumulusApiClient.invokeApifunction] - an async function to
+ * invoke the API Lambda that takes a prefix / user payload
+ * @param {integer} [params.timeout=30] - the number of seconds to wait for the
+ *   execution to reach a terminal state
+ * @returns {Promise<Object>} the execution as returned by the `GET /executions/<execution-arn>`
+ * endpoint
+ *
+ * @alias module:Executions
  */
-const getCompletedExecution = async (params = {}) =>
+const getCompletedExecution = async (params) =>
   pRetry(
     async () => {
       let execution;
@@ -89,17 +100,19 @@ const getCompletedExecution = async (params = {}) =>
   );
 
 /**
- * Wait for an execution to be failed and return it
+ * Wait for an execution status to be `failed` and return the execution
  *
  * @param {Object} params
  * @param {string} params.prefix - the prefix configured for the stack
  * @param {string} params.arn - an execution ARN
- * @param {Function} params.callback - an async function to invoke the api
- *   lambda that takes a prefix / user payload. Defaults to
- *   cumulusApiClient.invokeApifunction to invoke the api lambda
- * @param {integer} params.timeout - the number of seconds to wait for the
- *   execution to reach a terminal state. Defaults to 30.
- * @returns {Promise<Object>} the failed execution
+ * @param {Function} [params.callback=cumulusApiClient.invokeApifunction] - an async function to
+ * invoke the API Lambda that takes a prefix / user payload
+ * @param {integer} [params.timeout=30] - the number of seconds to wait for the
+ *   execution to reach a terminal state
+ * @returns {Promise<Object>} the execution as returned by the `GET /executions/<execution-arn>`
+ * endpoint
+ *
+ * @alias module:Executions
  */
 const getFailedExecution = async (params = {}) =>
   pRetry(
