+++ packages/integration-tests/package.json
@@ -19,6 +19,7 @@
   },
   "scripts": {
     "build": "rm -rf dist && mkdir dist && webpack",
+    "build-docs": "jsdoc2md --template README.hbs --param-list-format list Collections.js Executions.js Granules.js Providers.js Rules.js > README.md",
     "lint": "eslint .",
     "prepare": "npm run build"
   },
@@ -53,6 +54,7 @@
     "babel-loader": "^8.0.6",
     "babel-plugin-source-map-support": "^2.1.1",
     "babel-preset-env": "^1.7.0",
+    "jsdoc-to-markdown": "^5.0.3",
     "webpack": "~4.5.0",
     "webpack-cli": "~2.0.14"
   }
